Created Week card and moved all label content, functions and state for modal.
